requirements.txt
async-timeout==4.0.2
bitstruct==8.15.1
cc1101-python==1.2.1
ifaddr==0.2.0
ioctl-opt==1.2.2
paho-mqtt==1.6.1
zeroconf==0.39.1
